+++ test/libsolidity/smtCheckerTests/operators/unary_add_array_push_2.sol
@@ -1,0 +1,11 @@
+pragma experimental SMTChecker;
+
+contract C {
+	struct S {
+		int[][] d;
+	}
+	S[] data;
+	function f() public {
+		++data[1].d[3].push();
+	}
+}